fixes from pristine redo

+++ setenv.sh
@@ -19,6 +19,15 @@ export DEPOT_TOOLS_BOOTSTRAP_PYTHON3=0
 gclient metrics --opt-out
 
 # remove any previous attempts of fetch
-# so as to pull a fresh copy
-rm ${HOME}/zopen/dev/v8base/.gclient_entries
-rm ${HOME}/zopen/dev/v8base/.gclient
+# so as to pull a fresh copy.  If this is the first time building
+# depot tools, these file will not be present and thus
+# no need to attempt to remove them. 
+# NOTE: assuming v8 will be built in $HOME/zopen/dev/v8base
+if [ -d "${HOME}/zopen/dev/v8base" ]; then
+    if [ -f "${HOME}/zopen/dev/v8base/.gclient_entries" ]; then
+        rm ${HOME}/zopen/dev/v8base/.gclient_entries
+    fi
+    if [ -f "${HOME}/zopen/dev/v8base/.gclient_entries" ]; then
+        rm ${HOME}/zopen/dev/v8base/.gclient
+    fi
+fi
